pyproject.toml
[tool.poetry]
name = "thread_title_generator"
version = "0.1.0"
description = ""
authors = ["ykicisk <ykic.p3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
jupyterlab = "^2.2.8"
lxml = "^4.5.2"
emoji = "^0.6.0"
tensorflow-gpu = "^2.4.0"
tensorflow-datasets = "^3.2.1"
sklearn = "^0.0"
sentencepiece = "^0.1.94"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
